auth_service/requirements.txt
fastapi
uvicorn
bcrypt
pyjwt
python-dotenv
python-multipart
sqlalchemy
psycopg2-binary